Split aggregator command lines in a dedicated module
